+++ src/main/java/com/nickesqueda/socialmediademo/repository/UserRepository.java
@@ -5,12 +5,12 @@ import com.nickesqueda.socialmediademo.exception.ResourceNotFoundException;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<UserEntity, Integer> {
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
   Optional<UserEntity> findByUsername(String username);
 
   boolean existsByUsername(String username);
 
-  default UserEntity retrieveOrElseThrow(int userId) {
+  default UserEntity retrieveOrElseThrow(Long userId) {
     return findById(userId)
         .orElseThrow(() -> new ResourceNotFoundException(UserEntity.class, userId));
   }
